feat: compacted attachment code
